--- configs/test-utils/src/CustomError.ts
@@ -1,6 +1,0 @@
-export class CustomError extends Error {
-  constructor(...args: ConstructorParameters<ErrorConstructor>) {
-    super(...args)
-    console.error(...args)
-  }
-}
