--- configs/test-utils/src/index.ts
@@ -1,8 +1,0 @@
-export { CustomError } from './CustomError'
-export { CustomNotError } from './CustomNotError'
-export { Suspend } from './Suspend'
-export { ThrowError, ThrowNull } from './ThrowError'
-export { sleep } from './sleep'
-export const TEXT = 'TEXT' as const
-export const ERROR_MESSAGE = 'ERROR_MESSAGE' as const
-export const FALLBACK = 'FALLBACK' as const
