--- configs/test-utils/src/sleep.ts
@@ -1,1 +1,0 @@
-export const sleep = (ms: number) => new Promise<undefined>((resolve) => setTimeout(() => resolve(undefined), ms))
